refactor(samples): first pass at splitting topics.js into separate samples
Work in progress for this issue:
https://github.com/googleapis/nodejs-pubsub/issues/733
